Changed startCapture function to take an int instead of a long for number of frames.  If we need more than 2 billion frames taken, we should probably just do continuous capture.

+++ core/trunk/src/edu/cmu/ri/crw/AsyncVehicleServer.java
@@ -27,7 +27,7 @@ public interface AsyncVehicleServer {
         
         public void addCameraListener(CameraListener l, FunctionObserver<Void> obs);
 	public void removeCameraListener(CameraListener l, FunctionObserver<Void> obs);
-	public void startCamera(long numFrames, double interval, int width, int height, FunctionObserver<Void> obs);
+	public void startCamera(int numFrames, double interval, int width, int height, FunctionObserver<Void> obs);
 	public void stopCamera(FunctionObserver<Void> obs);
 	public void getCameraStatus(FunctionObserver<CameraState> obs);
